+++ packages/core/src/utils/create-substream.ts
@@ -1,9 +1,10 @@
-import { Package } from "../proto.js";
+import { fromBinary } from "@bufbuild/protobuf";
+import { type Package, PackageSchema } from "../proto.js";
 
-export function createSubstream(value: Uint8Array | ArrayBuffer) {
+export function createSubstream(value: Uint8Array | ArrayBuffer): Package {
   if (value instanceof ArrayBuffer) {
-    return Package.fromBinary(new Uint8Array(value));
+    return fromBinary(PackageSchema, new Uint8Array(value));
   }
 
-  return Package.fromBinary(value);
+  return fromBinary(PackageSchema, value);
 }