+++ packages/core/src/utils/get-output-type.ts
@@ -1,9 +1,9 @@
-import type { IMessageTypeRegistry } from "@bufbuild/protobuf";
+import type { Registry } from "@bufbuild/protobuf";
 import type { Module } from "../proto.js";
 import { getProtoType } from "./get-proto-type.js";
 import { isMapModule } from "./is-map-module.js";
 
-export function getOutputType(module: Module, registry: IMessageTypeRegistry) {
+export function getOutputType(module: Module, registry: Registry) {
   if (!isMapModule(module)) {
     return undefined;
   }