+++ packages/core/src/utils/get-proto-type.ts
@@ -1,18 +1,18 @@
-import type { IMessageTypeRegistry } from "@bufbuild/protobuf";
+import type { Registry } from "@bufbuild/protobuf";
 import { getProtoTypeName } from "./get-proto-type-name.js";
 
-export function getProtoType(typeName: string, registry: IMessageTypeRegistry) {
+export function getProtoType(typeName: string, registry: Registry) {
   const protoTypeName = getProtoTypeName(typeName);
   if (protoTypeName === undefined) {
     return undefined;
   }
 
-  return registry.findMessage(protoTypeName);
+  return registry.getMessage(protoTypeName);
 }
 
 export function getProtoTypeOrThrow(
   typeName: string,
-  registry: IMessageTypeRegistry,
+  registry: Registry,
   message = `Type "${typeName}" not found in registry`,
 ) {
   const type = getProtoType(typeName, registry);