+++ packages/core/src/utils/store-mode-name.ts
@@ -1,10 +1,9 @@
-import { proto3 } from "@bufbuild/protobuf";
-import { Module_Input_Store_Mode } from "../proto.js";
+import { Module_Input_Store_Mode, Module_Input_Store_ModeSchema } from "../proto.js";
 
 export function storeModeName(mode: Module_Input_Store_Mode): string {
-  const info = proto3.getEnumType(Module_Input_Store_Mode).findNumber(mode);
+  const info = Module_Input_Store_ModeSchema.value[mode];
   if (info === undefined) {
-    throw mode.toString().toLowerCase();
+    return mode.toString().toLowerCase();
   }
 
   return info.name.toLowerCase();