+++ packages/core/src/utils/unpack-map-output.ts
@@ -1,12 +1,13 @@
-import type { IMessageTypeRegistry } from "@bufbuild/protobuf";
+import { type Registry } from "@bufbuild/protobuf";
+import { anyUnpack } from "@bufbuild/protobuf/wkt";
 import type { Response } from "../proto.js";
 
-export function unpackMapOutput(response: Response, registry: IMessageTypeRegistry) {
+export function unpackMapOutput(response: Response, registry: Registry) {
   if (response.message.case === "blockScopedData") {
     const output = response.message.value.output?.mapOutput;
 
     if (output !== undefined) {
-      const message = output.unpack(registry);
+      const message = anyUnpack(output, registry);
       if (message === undefined) {
         throw new Error(`Failed to unpack output of type ${output.typeUrl}`);
       }